Split remaining Kubernetes imports from agent
* Move mesh configmap watcher out of the core `mesh` package
* Move security kubernetes `util`s to its own package

This drops the Kubernetes dependency for the agent completely.

This essentially completes https://github.com/istio/istio/issues/26232
for the agent (but not Istiod).
