+++ security/pkg/k8s/configutil_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package util
+package k8s
 
 import (
 	"context"
